backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-mail==1.4.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# 文档解析
python-docx==1.1.0
openpyxl==3.1.2

# RAG & LLM
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.8
langchain-google-genai==1.0.5
zhipuai>=2.1.0
chromadb==0.5.0
sentence-transformers==3.0.1

# WebSocket
websockets==12.0

# 工具
aiofiles==23.2.1
httpx==0.27.0

# 网页爬虫
beautifulsoup4>=4.12.0
lxml>=5.0.0
